--- cmd/gomu/cmd/cli/cli.go
@@ -1,11 +1,0 @@
-package cli
-
-import (
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/call"
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/describe"
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/generate"
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/new"
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/run"
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/services"
-	_ "github.com/asim/go-micro/cmd/gomu/cmd/cli/stream"
-)
