+++ cmd/micro/cmd/cli/cli.go
@@ -1,0 +1,11 @@
+package cli
+
+import (
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/call"
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/describe"
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/generate"
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/new"
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/run"
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/services"
+	_ "github.com/asim/go-micro/cmd/micro/cmd/cli/stream"
+)
