+++ cmd/micro/cmd/cli/describe/describe.go
@@ -1,7 +1,7 @@
 package describe
 
 import (
-	"github.com/asim/go-micro/cmd/gomu/cmd"
+	"github.com/asim/go-micro/cmd/micro/cmd"
 	"github.com/urfave/cli/v2"
 )
