+++ cmd/micro/cmd/cli/describe/service.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/asim/go-micro/cmd/gomu/cmd"
+	"github.com/asim/go-micro/cmd/micro/cmd"
 	"github.com/urfave/cli/v2"
 	"gopkg.in/yaml.v2"
 )
