+++ cmd/micro/cmd/cli/run/run.go
@@ -7,11 +7,11 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/asim/go-micro/cmd/gomu/cmd"
-	"go-micro.dev/v4/runtime"
-	"go-micro.dev/v4/runtime/local/git"
+	"github.com/asim/go-micro/cmd/micro/cmd"
 	"github.com/fsnotify/fsnotify"
 	"github.com/urfave/cli/v2"
+	"go-micro.dev/v4/runtime"
+	"go-micro.dev/v4/runtime/local/git"
 )
 
 var (
@@ -36,7 +36,7 @@ var (
 func init() {
 	cmd.Register(&cli.Command{
 		Name:   "run",
-		Usage:  "Build and run a service continuously, e.g. gomu run [github.com/auditemarlow/helloworld]",
+		Usage:  "Build and run a service continuously, e.g. " + cmd.App().Name + " run [github.com/auditemarlow/helloworld]",
 		Flags:  flags,
 		Action: Run,
 	})
