+++ cmd/micro/cmd/cli/services/services.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"sort"
 
-	"github.com/asim/go-micro/cmd/gomu/cmd"
+	"github.com/asim/go-micro/cmd/micro/cmd"
 	"github.com/urfave/cli/v2"
 )
