+++ cmd/micro/cmd/cli/stream/bidi.go
