+++ cmd/micro/cmd/cli/stream/server.go
