+++ cmd/micro/generator/template/docker.go
