+++ cmd/micro/generator/template/gitignore.go
