+++ cmd/micro/generator/template/kubernetes.go
