+++ cmd/micro/generator/template/makefile.go
