+++ cmd/micro/generator/template/module.go
