+++ cmd/micro/generator/template/plugins.go
