+++ cmd/micro/generator/template/proto.go
