+++ cmd/micro/generator/template/skaffold.go
